Balance GD

+++ Assets/_Scripts/Weapons/WeaponStatisticUgradeSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "WeaponStatisticUgrade", menuName = "ScriptableObjects/Weapons/WeaponStatisticUgrade")]
 public class WeaponStatisticUgradeSO : ScriptableObject
 {
-    public int Damages;
+    public float Damages;
     public float AttackSpeed;
     public float Range;
 }
